requirements.txt

termcolor
pandas
cassandra-driver
